refactor: Clear localStorage and remove unused code in cache.js
