+++ AIBlogFrontend/src/pages/AdminDashboard/index.jsx
@@ -1,5 +1,6 @@
-import CreatePost from "./CreatePost";
-import EditPost from "./EditPost";
+import { CreatePost } from "./CreatePost";
+import { ListPosts } from "./ListPosts";
+import { Logout } from "./components/Logout";
 
 const AdminDashboard = () => {
   if (localStorage.getItem("token") != import.meta.env.VITE_CUSTOM_TOKEN) {
@@ -8,10 +9,18 @@ const AdminDashboard = () => {
 
   return (
     <div className="container mx-auto py-8">
-      <h1 className="text-3xl font-bold text-gray-800">Admin Dashboard</h1>
-      <div className="flex md:flex-col sm:flex-wrap gap-4 mt-4">
-        <CreatePost />
-        <EditPost />
+      <div className="flex justify-between items-center">
+        <h1 className="text-3xl font-bold text-gray-800">Admin Dashboard</h1>
+        <Logout />
+      </div>
+
+      <div className="flex flex-col md:flex-row md:flex-wrap gap-4 mt-4">
+        <div className="w-full md:w-1/2 mb-0 sm:mb-10">
+          <CreatePost />
+        </div>
+        <div className="w-full md:w-1/2">
+          <ListPosts />
+        </div>
       </div>
     </div>
   );
