+++ AIBlogFrontend/src/services/util.js
@@ -25,13 +25,3 @@ export const getFooterHeight = () => {
 //     return null;
 //   }
 // };
-
-export const categories = [
-  {
-    name: "Understanding Concepts in AI",
-    value: "Understanding Concepts in AI",
-  },
-  { name: "AI in education", value: "AI in education" },
-  { name: "AI in News/Trends", value: "AI in News/Trends" },
-  { name: "Safety and Ethics of AI", value: "Safety and Ethics of AI" },
-];
